Update test script

+++ test/_test.sh
@@ -5,15 +5,15 @@
 
 set -euxo pipefail
 
-: ${BIN_LOC:="./target/debug/qstract"}
-QSTRACT_BIN="$(realpath $BIN_LOC)"
+: "${BIN_LOC:="./target/debug/qstract"}"
+QSTRACT_BIN="$(realpath "$BIN_LOC")"
 TEST_DIR="$(realpath ./test)"
 
 TEMP_DIR="$(mktemp -d)"
 pushd "$TEMP_DIR"
 
 # Copy all test files
-cp -r $TEST_DIR/* .
+cp -r "$TEST_DIR"/* .
 ls .
 
 # Helpers
@@ -69,13 +69,13 @@ function test_1 {
 
 function test_2 {
     $QSTRACT_BIN -C ./em1 -z ./e1.tar.gz
-    if [ ! -z "$(ls -A ./em1)" ]; then
+    if [ -n "$(ls -A ./em1)" ]; then
         echo "Not empty"
         exit 1
     fi
 
     $QSTRACT_BIN -C ./em2 ./e1.tar
-    if [ ! -z "$(ls -A ./em2)" ]; then
+    if [ -n "$(ls -A ./em2)" ]; then
         echo "Not empty"
         exit 1
     fi
@@ -94,7 +94,7 @@ function test_4 {
 
     $QSTRACT_BIN ./zarchive1.zip -C ./ --zip
     10_files ./
-    rm -rf *.txt
+    rm -rf ./*.txt
 
     $QSTRACT_BIN -C ./t555/111 --zip ./zarchive1.zip
     10_files ./t555/111
@@ -108,7 +108,7 @@ function test_5 {
 
     $QSTRACT_BIN ./7z-zarchive1.zip -C ./ --zip
     10_files ./
-    rm -rf *.txt
+    rm -rf ./*.txt
 
     $QSTRACT_BIN -C ./t555/111 --zip ./7z-zarchive1.zip
     10_files ./t555/111
@@ -126,27 +126,31 @@ function test_6 {
 }
 
 function test_7 {
-    $QSTRACT_BIN --sha256 ./tarchive1.tar.gz | grep 'dfd5fe115d931b6c4e4860391fc3984ab4ec5b6110851c0ef5a8641bd079f2c2' &>/dev/null
-    [ $? != 0 ] && exit 2
+    if ! $QSTRACT_BIN --sha256 ./tarchive1.tar.gz | grep 'dfd5fe115d931b6c4e4860391fc3984ab4ec5b6110851c0ef5a8641bd079f2c2' &>/dev/null; then
+        exit 2
+    fi
 
-    $QSTRACT_BIN ./tarchive1.tar.gz --sha512 | grep 'a73f29153d74694323e8d0bb82a910018a05ce011c73dfdca62b9fca3aafc218b92f2ff277a4b8006a919874527843d2ba0d56c0d5532487ac0a1d84421b5539' &>/dev/null
-    [ $? != 0 ] && exit 2
+    if ! $QSTRACT_BIN ./tarchive1.tar.gz --sha512 | grep 'a73f29153d74694323e8d0bb82a910018a05ce011c73dfdca62b9fca3aafc218b92f2ff277a4b8006a919874527843d2ba0d56c0d5532487ac0a1d84421b5539' &>/dev/null; then
+        exit 2
+    fi
 
-    $QSTRACT_BIN ./tarchive1.tar.gz --sha3_256 | grep '60e99497952e17852d63c843606a214945550ccbb6777db25f7405fb39b55025' &>/dev/null
-    [ $? != 0 ] && exit 2
+    if ! $QSTRACT_BIN ./tarchive1.tar.gz --sha3_256 | grep '60e99497952e17852d63c843606a214945550ccbb6777db25f7405fb39b55025' &>/dev/null; then
+        exit 2
+    fi
 
-    $QSTRACT_BIN --sha3_512 ./tarchive1.tar.gz | grep '2b292cf7c24d023e5f4b836ff36640b1a6b40d94382feab658b7c5ac5aa65ba7bdad2756e2c1c3077dd446b4474bdad72e96a030d0a291408482ac668246e4a' &>/dev/null
-    [ $? != 0 ] && exit 2
+    if ! $QSTRACT_BIN --sha3_512 ./tarchive1.tar.gz | grep '2b292cf7c24d023e5f4b836ff36640b1a6b40d94382feab658b7c5ac5aa65ba7bdad2756e2c1c3077dd446b4474bdad72e96a030d0a291408482ac668246e4a' &>/dev/null; then
+        exit 2
+    fi
 
-    if [ "$($QSTRACT_BIN --sha512 -z ./e1.tar)" ]; then
+    if $QSTRACT_BIN --sha512 -z ./e1.tar; then
         exit 1
     fi
 
-    if [ "$($QSTRACT_BIN --sha256 --sha512 ./e1.tar)" ]; then
+    if $QSTRACT_BIN --sha256 --sha512 ./e1.tar; then
         exit 1
     fi
 
-    if [ "$($QSTRACT_BIN --wrong ./e1.tar)" ]; then
+    if $QSTRACT_BIN --wrong ./e1.tar; then
         exit 1
     fi
 }
@@ -162,4 +166,4 @@ test_6
 test_7
 
 popd
-rm -rf "$TEMP_DIR"
+[[ "$TEMP_DIR" != "" ]] && rm -rf "$TEMP_DIR"
